+++ tools/generator/model/src/main/java/is/codion/tools/generator/model/DomainGeneratorModel.java
@@ -464,7 +464,7 @@ public final class DomainGeneratorModel {
 		}
 
 		@Override
-		public void onResult() {
+		public void onSuccess() {
 			schemaTableModel.selection().indexes().get().forEach(index ->
 							schemaTableModel.fireTableRowsUpdated(index, index));
 			schemaSelectionChanged();